Install "npm" in CentOS and Fedora
In RPM distributions like CentOS and Fedora the "npm" package is separate, and
not installed by default when Node.js is installed. As a result the execution
of the "bin/setup" will fail if the instructions are followed to setup the
environment from scratch, because it uses the "npm" command to install bower.
This patch changes the instructions so that the "npm" package is also
installed.

Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>