Add a note on missing developer tools in OS X
In case ManageIQ is being installed on a fresh OS X chances are that the
developer tools have not been installed. This results in an error about
missing symbols and/or libraries during the installation of Nokogiri gem.

The patch adds a brief explanation in the troubleshooting section
explaining how to install developer tools and try `bin/setup` again.

Signed-off-by: Gregor Berginc <gregor.berginc@xlab.si>