Update document on running test suites
Instructions in the document were a bit out-of-sync with the actual
state of affairs and this commit attempts to fix this.

There are two major things that we did in order to get documentation
back up to speed. First, we removed the parts of the documentation
that were no longer true and added some generic documentation on how
to run tests in vast majority of ManageIQ repos.

To finish off the sync operation, we expanded the section about the UI
tests that behave quite differently from the rest of the bunch.