[openshift.md] Fix list formatting
With the current implementation of this code, the indenting is
incorrect, and causes the generated html markup to generate a new list
element for each new list item.

The adding of the whitespace in this commit fixes the issue, and allows
the markdown processor to properly create a single list.