[.rubocop-base.yml] Disable new cops by default
We were presented with this warning as part of the upgrade process to
rubocop-0.82.0

    Linter::Rubocop STDERR:
    ```
    The following cops were added to RuboCop, but are not configured. Please set Enabled to either `true` or `false` in your `.rubocop.yml` file:
     - Layout/SpaceAroundMethodCallOperator (0.82)
     - Lint/RaiseException (0.81)
     - Lint/StructNewOverride (0.81)
     - Style/ExponentialNotation (0.82)
     - Style/HashEachMethods (0.80)
     - Style/HashTransformKeys (0.80)
     - Style/HashTransformValues (0.80)

    For more information: https://docs.rubocop.org/en/latest/versioning/
    ```

And the documentation describes the usage of `NewCops: disabled` as a
way to avoid this error message when upgrading rubocop.

Arguably this isn't wanted, but the intent is that rubocop upgrades in
the future won't cause rubocop to crash, and we can get to evaluating
our use of the new cops when time permits.